feat!: first anniversary and first stable release of dart packages (#102)
* feat(example): added attributes for profiling of example

* chore(deps): bump dependencies

* feat(l10n): added new language translations from a new source

* feat(cli): update cli tools for new syntax

* refactor(ide): removed dimmer and updated ide settings

* refactor(data): reorder singapore languages
https://en.wikipedia.org/wiki/Languages_of_Singapore#English_as_the_main_language_of_Singapore

* chore(deps): bump dependencies

* refactor: small refactoring before data update

* feat(l10n): added more rare language name translations

* feat(l10n): added german translations for language names

* refactor: update language translation tests and docs

* feat(model): added common getter for translated name

* style(docs): unify docs for currencies

* fix(lint): added line breaks to fit 80 chars limit

* feat(l10n): added new translations for currencies

* feat(cli): added php to json converter cli tool

* feat(ci): added dependabot for cli tools

* feat(l10n): provide material locale translations for all currencies
Closes: https://github.com/users/tsinis/projects/1?query=is%3Aopen+sort%3Aupdated-desc&pane=issue&itemId=42273773

* refactor(model)!: remove regular list, add list extended

* feat(l10n): added more translations, deprecated hrk

* feat(cli): added l10n from simple map in php converter

* fix(cli): added type check, in data list generator

* feat(l10n): additional translations for countries from a new source

* docs(license): update licenses with l10n authors

* refactor!: removed deprecated code

* docs: updated package versions and readmes

* docs: added changelogs for dart packages v1.0.0

* feat: added and referenced cupertino-supported locales list

+++ tools/bin/translations.dart
@@ -7,16 +7,20 @@ import "package:cli/utils/io_utils.dart";
 import "package:cli/utils/json_utils.dart";
 import "package:cli/utils/network_utils.dart";
 
-/// Usage: `dart run :translations sealed_currencies`.
-Future<void> main(List<String> args) async {
+/// Usage: `dart run :translations sealed_countries`.
+Future<void> main(List<String> args, {bool skipFetch = true}) async {
   final package = ArgsParser(args).maybePackageName();
   if (package == null) throw ArgumentError("Package name should be provided.");
+  final io = IoUtils();
 
-  final clonedDir =
-      await const NetworkUtils().cloneRepository(package.umpirskyRepoUrl);
-  final io = IoUtils()
-    ..moveJsonFiles(clonedDir)
-    ..deleteDirectory(clonedDir);
+  if (!skipFetch) {
+    final clonedDir =
+        await const NetworkUtils().cloneRepository(package.umpirskyRepoUrl);
+
+    io
+      ..moveJsonFiles(clonedDir)
+      ..deleteDirectory(clonedDir);
+  }
 
   final exports = await JsonUtils(package).parseByLanguage();
 
@@ -24,8 +28,10 @@ Future<void> main(List<String> args) async {
   final buffer = StringBuffer(
     """
 // This library translations are based on the data from the
-// ${package.umpirskyRepoUrl} project (from Saša Stamenković),
-// which is licensed under the MIT License.
+// https://github.com/symfony/intl project
+// (from The Symfony - Intl Component, Fabien Potencier) and from the
+// ${package.umpirskyRepoUrl} project (from Saša Stamenković).
+// Both projects are licensed under the MIT License.
 
 /// Provides $dataType translations for ${package.dirName}.
 """,
